+++ lib/fue/finder.rb
@@ -19,10 +19,11 @@ module Fue
 
     def emails(options = {})
       query = <<-GRAPHQL
-        query($login: String!, $author_id: ID!, $depth: Int!, $breadth: Int!) {
+        query($login: String!, $author_id: ID!, $depth: Int!, $breadth: Int!, $breadthCursor: String) {
           user(login: $login) {
-            repositories(last: $breadth, isFork:false, privacy: PUBLIC) {
+            repositories(last: $breadth, after: $breadthCursor, isFork:false, privacy: PUBLIC) {
               edges {
+                cursor
                 node {
                   defaultBranchRef {
                     target {
@@ -45,19 +46,35 @@ module Fue
         }
       GRAPHQL
 
+      max_breadth = options[:breadth] || 10
+
       query_options = {
         login: options[:username],
         author_id: author_id(options[:username]),
-        depth: options[:depth],
-        breadth: options[:breadth]
+        depth: options[:depth] || 1
       }
 
-      graphql_client.query(query, query_options).data.user.repositories.edges.map do |edge|
-        next unless edge.node.default_branch_ref
-        edge.node.default_branch_ref.target.history.nodes.map do |node|
-          "#{node.author.name} <#{node.author.email}>"
+      emails = Set.new
+
+      loop do
+        query_options[:breadth] = [max_breadth, 100].min
+        response = graphql_client.query(query, query_options)
+        edges = response.data.user.repositories.edges if response
+        if edges
+          edges.each do |edge|
+            query_options[:breadthCursor] = edge.cursor
+            branch_ref = edge.node.default_branch_ref
+            next unless branch_ref
+            branch_ref.target.history.nodes.each do |node|
+              emails << "#{node.author.name} <#{node.author.email}>"
+            end
+          end
         end
-      end.flatten.compact.uniq
+        max_breadth -= 100
+        break if max_breadth <= 0
+      end
+
+      emails.to_a
     end
 
     private